Fixed GC issues with write-and-pop redirection
Rewrote the redirect writeAndPop (literals, fields) to use newly introduced methods to create contexts directly and populate them with the correct arguments directly.
Does not rely on typical calling convention anymore.
Avoids reseting the newTop slot of the stack of the old frame. Resetting is necessary since some parts rely on the newTop slot having the right value, but resetting it after a standard send is complex, because of potential GC in-between, and it was ugly.

Signed-off-by: Stefan Marr <git@stefan-marr.de>